+++ 03week/TowersOfHanoi/readMe.md
@@ -1,0 +1,15 @@
+- don't check for a win on array a (turn into a test)
+
+- get our user input(startStack,endStack) and then check the rules of the game based on the current object
+
+- to move a # we need to update the array object using our user input that we get (startStack, endStack)
+
+- whenever we want to move a #, we need to check the #s and make it is less than the # we are trying to put it on.
+
+- to determine the game has been won, check that either array b or c (turn into a test) has all of the #s in ascending order (greatest to least)
+
+- should only be able to remove a # from the end of an array and place it at the end of another array.
+
+- if array is empty # are allowed to be moves their
+
+- if you choose am empty array nothing will happen and error message will be console logged out.
